+++ server/heroku/oauth.go
@@ -15,13 +15,19 @@ const (
 type Authorization heroku.OAuthAuthorization
 
 func newAuthorization(token *AccessToken) *Authorization {
+	var expIn *int
+	if t := token.ExpiresAt; t != nil {
+		exp := int(token.ExpiresIn().Seconds())
+		expIn = &exp
+	}
 	return &Authorization{
 		AccessToken: &struct {
 			ExpiresIn *int   `json:"expires_in"`
 			Id        string `json:"id"`
 			Token     string `json:"token"`
 		}{
-			Token: token.Token,
+			ExpiresIn: expIn,
+			Token:     token.Token,
 		},
 	}
 }