+++ server/server.go
@@ -6,8 +6,10 @@ package server
 import (
 	"io"
 	"net/http"
+	"net/url"
 
 	"github.com/remind101/empire"
+	"github.com/remind101/empire/pkg/saml"
 	"github.com/remind101/empire/server/github"
 	"github.com/remind101/empire/server/heroku"
 	"github.com/remind101/pkg/httpx"
@@ -35,14 +37,21 @@ type Options struct {
 // Server composes the Heroku API compatibility layer, the GitHub Webhooks
 // handlers and a health check as a single http.Handler.
 type Server struct {
+	// Base host for the server.
+	URL *url.URL
+
 	// The underlying Heroku http.Handler.
 	Heroku *heroku.Server
+
+	// If provided, enables the SAML integration.
+	ServiceProvider *saml.ServiceProvider
 
 	mux *httpx.Router
 }
 
 func New(e *empire.Empire, options Options) *Server {
 	r := httpx.NewRouter()
+	s := &Server{mux: r}
 
 	if options.GitHub.Webhooks.Secret != "" {
 		// Mount GitHub webhooks
@@ -55,16 +64,17 @@ func New(e *empire.Empire, options Options) *Server {
 	}
 
 	// Mount the heroku api
-	hk := heroku.New(e)
-	r.Headers("Accept", heroku.AcceptHeader).Handler(hk)
+	s.Heroku = heroku.New(e)
+	r.Headers("Accept", heroku.AcceptHeader).Handler(s.Heroku)
+
+	// Mount SAML handlers.
+	r.HandleFunc("/saml/login", s.SAMLLogin)
+	r.HandleFunc("/saml/acs", s.SAMLACS)
 
 	// Mount health endpoint
 	r.Handle("/health", NewHealthHandler(e))
 
-	return &Server{
-		Heroku: hk,
-		mux:    r,
-	}
+	return s
 }
 
 func (s *Server) ServeHTTPContext(ctx context.Context, w http.ResponseWriter, r *http.Request) error {