+++ tests/api/api_test.go
@@ -27,8 +27,14 @@ func NewTestClient(t testing.TB) (*heroku.Client, *empiretest.Server) {
 	e := empiretest.NewEmpire(t)
 	s := empiretest.NewServer(t, e)
 	s.Heroku.Auth = &auth.Auth{
-		Authenticator: auth.Anyone(&empire.User{Name: "fake"}),
+		Strategies: auth.Strategies{
+			{
+				Name:          auth.StrategyUsernamePassword,
+				Authenticator: auth.Anyone(&empire.User{Name: "fake"}),
+			},
+		},
 	}
+	s.Start()
 
 	c := &heroku.Client{
 		Username: "",